requirements.txt
absl-py==2.3.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alabaster==1.0.0
aniso8601==10.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asciitree==0.3.3
asttokens==3.0.0
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
backports-datetime-fromisoformat==2.0.3
beautifulsoup4==4.13.4
black==19.10b0
boto3==1.39.9
botocore==1.39.9
braceexpand==0.1.7
cdifflib==1.2.9
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.0.2
clip==0.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
cytoolz==1.0.1
datasets==2.19.0
decorator==5.2.1
diffusers==0.34.0
dill==0.3.8
Distance==0.1.3
docopt==0.6.2
docutils==0.21.2
editdistance==0.8.1
einops==0.8.1
einops-exts==0.0.4
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.11.0.post1
fasteners==0.19
fasttext==0.9.3
filelock==3.13.1
Flask==2.2.5
Flask-RESTful==0.3.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.3.1
ftfy==6.3.1
future==1.0.0
g2p-en==2.1.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.73.1
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.21.3
hydra-core==1.3.2
idna==3.10
ijson==3.4.0
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.7.0
inflect==7.5.0
iniconfig==2.1.0
intervaltree==3.1.0
ipython==8.37.0
ipywidgets==8.1.7
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.4
jiwer==2.5.2
jmespath==1.0.1
joblib==1.5.1
jupyterlab_widgets==3.0.15
kaldi-python-io==1.2.2
kaldiio==2.18.1
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
latexcodec==3.0.1
lazy_loader==0.4
Levenshtein==0.22.0
lhotse==1.30.3
librosa==0.11.0
lightning-utilities==0.14.3
lilcom==1.8.1
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
marshmallow==4.0.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
megatron_core==0.5.0
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
nemo_text_processing==1.1.0
nemo_toolkit
nerfacc==0.5.3
networkx==3.3
nltk==3.9.1
numba==0.61.2
numcodecs==0.13.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
onnx==1.18.0
open_clip_torch==2.32.0
OpenCC==1.1.6
packaging==25.0
pandas==2.3.1
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
plac==1.4.5
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
portalocker==3.2.0
progress==1.6.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyarrow==21.0.0
pyarrow-hotfix==0.7
pybind11==3.0.0
pybtex==0.25.1
pybtex-docutils==1.0.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyloudnorm==0.1.1
PyMCubes==0.1.6
pynini==2.1.6.post1
pyparsing==3.2.3
pypinyin==0.55.0
pypinyin-dict==0.9.0
PySocks==1.7.1
pytest==8.4.1
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.0.7
pytz==2025.2
PyYAML==6.0.2
rapidfuzz==2.13.7
regex==2024.11.6
requests==2.32.4
resampy==0.4.3
rich==14.0.0
rouge_score==0.1.2
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.1
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.0.0
sentencepiece==0.2.0
sentry-sdk==2.33.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
sox==1.5.0
soxr==0.5.0.post1
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.6.5
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
taming-transformers==0.0.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorstore==0.1.76
termcolor==3.1.0
text-unidecode==1.3
textdistance==4.6.3
texterrors==1.0.9
threadpoolctl==3.6.0
timm==1.0.17
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchdiffeq==0.2.5
torchmetrics==1.7.4
torchsde==0.2.6
torchvision==0.20.1+cu118
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.40.0
trimesh==4.7.1
triton==3.1.0
typed-ast==1.5.5
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.0
wcwidth==0.2.13
webdataset==0.1.62
Werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.14
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
youtokentome==1.0.6
zarr==2.18.3
zipp==3.23.0
